+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG.md
 
+## 0.0.4 (2024-09-25)
+- Added `rbxRankDB:getRankRange` method to get a range of elements from a list based on rank.
+
 ## 0.0.3 (2024-09-25)
-- Add methods to update/retrieve multiple elements in one request. `rbxRankDB:updateMultiElements` and `rbxRankDB:getMultiElements`
+- Added methods to update/retrieve multiple elements in one request. `rbxRankDB:updateMultiElements` and `rbxRankDB:getMultiElements`
 
 ## 0.0.2 (2024-09-24)